requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Machine Learning - ACTIVATED
xgboost==2.0.3
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Redis Cache
redis==5.0.1
hiredis==2.2.3

# API Clients
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Market Data
alpha-vantage==2.3.1
polygon-api-client==1.12.4
yfinance==0.2.32

# Websockets
websockets==12.0

# Task Scheduling
apscheduler==3.10.4

# Email
aiosmtplib==3.0.1

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
python-dotenv==1.0.0
